2010-08-09	Bharath Suri <bharath.s.jois@gmail.com>
	PR 1660/filesystem
	* libfs/src/pipe/fifo.c, libfs/src/pipe/pipe.h: Changed
	int pipe_release to void pipe_release.
	* libfs/src/imfs/imfs_fifo.c: Corresponding change to
	IMFS_fifo_close since pipe_release does not return any error.
	* libfs/src/imfs/imfs_initsupp.c: Changes to improve
	IMFS_determine_bytes_per_block

+++ cpukit/score/include/rtems/score/object.h
@@ -476,7 +476,7 @@ void _Objects_Shrink_information(
 void _Objects_Initialize_information (
   Objects_Information *information,
   Objects_APIs         the_api,
-  uint32_t             the_class,
+  uint16_t             the_class,
   uint32_t             maximum,
   uint16_t             size,
   bool                 is_string,
@@ -754,7 +754,7 @@ Objects_Control *_Objects_Get_next(
  */
 Objects_Information *_Objects_Get_information(
   Objects_APIs   the_api,
-  uint32_t       the_class
+  uint16_t       the_class
 );
 
 /**
