+++ cpukit/score/src/objectinitializeinformation.c
@@ -49,7 +49,7 @@
 void _Objects_Initialize_information(
   Objects_Information *information,
   Objects_APIs         the_api,
-  uint32_t             the_class,
+  uint16_t             the_class,
   uint32_t             maximum,
   uint16_t             size,
   bool                 is_string,
